# core/generated/swagger-ui/schema-registry-api-spec.yaml
...
swagger: "2.0"
info:
  version: "v1"
  title: "Confluent Schema Registry"
schemes:
- "http"
- "https"
paths:
...
  /config:
    get:
      summary: "Get global compatibility level."
      description: ""
      operationId: "getTopLevelConfig"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Config"
        500:
          description: "Error code 50001 -- Error in the backend data store"
    put:
      summary: "Update global compatibility level."
      description: ""
      operationId: "updateTopLevelConfig"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - in: "body"
        name: "body"
        description: "Config Update Request"
        required: true
        schema:
          $ref: "#/definitions/ConfigUpdateRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ConfigUpdateRequest"
        422:
          description: "Error code 42203 -- Invalid compatibility level"
        500:
          description: "Error code 50001 -- Error in the backend data store\nError\
            \ code 50003 -- Error while forwarding the request to the primary\n"
  /config/{subject}:
    get:
      summary: "Get compatibility level for a subject."
      description: ""
      operationId: "getSubjectLevelConfig"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "subject"
        in: "path"
        required: true
        type: "string"
      - name: "defaultToGlobal"
        in: "query"
        required: false
        type: "boolean"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Config"
        404:
          description: "Subject not found"
        500:
          description: "Error code 50001 -- Error in the backend data store"
    put:
      summary: "Update compatibility level for the specified subject."
      description: ""
      operationId: "updateSubjectLevelConfig"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "subject"
        in: "path"
        description: "Name of the Subject"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Config Update Request"
        required: true
        schema:
          $ref: "#/definitions/ConfigUpdateRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ConfigUpdateRequest"
        422:
          description: "Error code 42203 -- Invalid compatibility level\nError code\
            \ 40402 -- Version not found"
        500:
          description: "Error code 50001 -- Error in the backend data store\nError\
            \ code 50003 -- Error while forwarding the request to the primary"
...
definitions:
...
  Config:
    type: "object"
    properties:
      compatibilityLevel:
        type: "string"
        description: "Compatability Level"
        enum:
        - "BACKWARD"
        - "BACKWARD_TRANSITIVE"
        - "FORWARD"
        - "FORWARD_TRANSITIVE"
        - "FULL"
        - "FULL_TRANSITIVE"
        - "NONE"
  ConfigUpdateRequest:
    type: "object"
    properties:
      compatibility:
        type: "string"
        description: "Compatability Level"
        enum:
        - "BACKWARD"
        - "BACKWARD_TRANSITIVE"
        - "FORWARD"
        - "FORWARD_TRANSITIVE"
        - "FULL"
        - "FULL_TRANSITIVE"
        - "NONE"
...